core/requirements.txt
django >4, <5
python-decouple
Pillow
djangorestframework
markdown      
django-filter
drf-yasg[validation]
